TASK: Remove aspect and make routes mandatory
The FrontendNodeRoutePartHandlerAspect caused some trouble because it is called whenever the
FrontendNodeRoutePartHandler->convertRequestPathToNode was called which can occurred in
multiple routes and cause various problems.

This change removes the aspect entirely and updates the documentation to make adding the package
routes at the end of the global routes.yaml mandatory.